+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Preserve image format in PIL.Image.Image when loading
 * Added ONNX version of Detectron2 and make default model
 * Remove API code, we don't serve this as a standalone API any more
+* Update ordering logic to account for multicolumn documents.
 
 ## 0.4.4